+++ board/recalbox/x86/fsoverlay/usr/bin/externalscreen.sh
@@ -4,12 +4,12 @@ export DISPLAY=:0
 
 xrandrOutput=$(xrandr)
 
-OUTPUT_MAXRES=$(echo "${xrandrOutput}" | awk '$1!~"[0-9]*x[0-9]*i?" {OUTPUT=$1; RESO=1;} $1~"[0-9]*x[0-9]*i?" && RESO==1 {print OUTPUT "=" $1; RESO++;}')
+OUTPUT_MAXRES=$(echo "${xrandrOutput}" | awk '$1!~"[0-9]*x[0-9]*i?" {if (MAXRES!="") {print OUTPUT ";" MAXRES ";" FALLBACKRES;} OUTPUT=$1; MAXRES=""; FALLBACKRES=""; NBRESO=1} $1~"[0-9]*x[0-9]*i?" {if (NBRESO==1) {MAXRES=$1; NBRESO++;} HEIGHT=$1; gsub(/^[0-9]+x/,"",HEIGHT); if ((FALLBACKRES=="") && (HEIGHT+0<=1080)) {FALLBACKRES=$1;}} END {if (MAXRES!="") {print OUTPUT ";" MAXRES ";" FALLBACKRES;}}')
 if [ -z "$OUTPUT_MAXRES" ]; then
   echo "No screen connected..."
   exit
 else
-  echo "Connected screen(s):"
+  echo "Connected screen(s) / Max resolution / Fallback resolution lower than FullHD:"
   for screen in ${OUTPUT_MAXRES[*]}; do
     echo "  ${screen}"
   done
@@ -21,12 +21,12 @@ FORCE_RES=$(recalbox_settings -command load -key system.externalscreen.forcereso
 FORCE_FREQ=$(recalbox_settings -command load -key system.externalscreen.forcefrequency)
 
 if [ -n "$PREFERED" ]; then
-  PREFERED_CONNECTED=$(echo "$OUTPUT_MAXRES" | grep "$PREFERED" | awk 'BEGIN {FS="="} {print $1}')
+  PREFERED_CONNECTED=$(echo "$OUTPUT_MAXRES" | grep "$PREFERED" | awk 'BEGIN {FS=";"} {print $1}')
 fi
 if [[ -n "$PREFERED" ]] && [[ -n "$PREFERED_CONNECTED" ]]; then
   SEL_OUTPUT="$PREFERED"
 else
-  SEL_OUTPUT=$(echo "$OUTPUT_MAXRES" | awk 'BEGIN {FS="="} NR==1 {print $1}')
+  SEL_OUTPUT=$(echo "$OUTPUT_MAXRES" | awk 'BEGIN {FS=";"} NR==1 {print $1}')
 fi
 if [ -n "$PREFERED" ]; then
   if [ -n "$PREFERED_CONNECTED" ]; then
@@ -44,26 +44,26 @@ if [ -n "$FORCE_RES" ]; then
       echo "Force selected output $SEL_OUTPUT to $FORCE_RES at frequence $FORCE_FREQ, disable others"
     else
       echo "Force selected output $SEL_OUTPUT to $FORCE_RES, disable others"
-  fi
-  XRANDR_CMD=$(echo "$OUTPUT_MAXRES" | awk -v SEL_OUTPUT="$SEL_OUTPUT" -v FORCE_RES="$FORCE_RES" -v FORCE_FREQ_OPTION="$FORCE_FREQ_OPTION" 'BEGIN {FS="="; printf "xrandr"} $1==SEL_OUTPUT {printf " --output " $1 " --mode " FORCE_RES FORCE_FREQ_OPTION;} $1!=SEL_OUTPUT {printf " --output " $1 " --off";} END {printf "\n";}') 
+    fi
+    XRANDR_CMD=$(echo "$OUTPUT_MAXRES" | awk -v SEL_OUTPUT="$SEL_OUTPUT" -v FORCE_RES="$FORCE_RES" -v FORCE_FREQ_OPTION="$FORCE_FREQ_OPTION" 'BEGIN {FS=";"; printf "xrandr"} $1==SEL_OUTPUT {printf " --output " $1 " --mode " FORCE_RES FORCE_FREQ_OPTION;} $1!=SEL_OUTPUT {printf " --output " $1 " --off";} END {printf "\n";}')
   else
-    SEL_OUTPUT_VERT_PIX=$(echo "$OUTPUT_MAXRES" | grep "$SEL_OUTPUT" | sed 's/^.*x\([0-9]*\)$/\1/')
-  if [ "$SEL_OUTPUT_VERT_PIX" -gt "1080" ]; then
-    echo "Force resolution inexistant for selected output $SEL_OUTPUT, fallback to FHD, disable other outputs"
-    XRANDR_CMD=$(echo "$OUTPUT_MAXRES" | awk -v SEL_OUTPUT="$SEL_OUTPUT" 'BEGIN {FS="="; printf "xrandr"} $1==SEL_OUTPUT {printf " --output " $1 " --mode 1920x1080";} $1!=SEL_OUTPUT {printf " --output " $1 " --off";} END {printf "\n";}')
-  else
+    SEL_OUTPUT_VERT_PIX=$(echo "$OUTPUT_MAXRES" | grep "$SEL_OUTPUT" | sed 's/^.*x\([0-9]*\);.*$/\1/')
+    if [ "$SEL_OUTPUT_VERT_PIX" -gt "1080" ]; then
+      echo "Force resolution inexistant for selected output $SEL_OUTPUT at fallback resolution, disable other outputs"
+      XRANDR_CMD=$(echo "$OUTPUT_MAXRES" | awk -v SEL_OUTPUT="$SEL_OUTPUT" 'BEGIN {FS=";"; printf "xrandr"} $1==SEL_OUTPUT {printf " --output " $1 " --mode " $3;} $1!=SEL_OUTPUT {printf " --output " $1 " --off";} END {printf "\n";}')
+    else
       echo "Force resolution inexistant for selected output $SEL_OUTPUT, fallback to auto, disable other outputs"
-    XRANDR_CMD=$(echo "$OUTPUT_MAXRES" | awk -v SEL_OUTPUT="$SEL_OUTPUT" 'BEGIN {FS="="; printf "xrandr"} $1==SEL_OUTPUT {printf " --output " $1 " --auto";} $1!=SEL_OUTPUT {printf " --output " $1 " --off";} END {printf "\n";}')
-  fi
+      XRANDR_CMD=$(echo "$OUTPUT_MAXRES" | awk -v SEL_OUTPUT="$SEL_OUTPUT" 'BEGIN {FS=";"; printf "xrandr"} $1==SEL_OUTPUT {printf " --output " $1 " --auto";} $1!=SEL_OUTPUT {printf " --output " $1 " --off";} END {printf "\n";}')
+    fi
   fi
 else
-  SEL_OUTPUT_VERT_PIX=$(echo "$OUTPUT_MAXRES" | grep "$SEL_OUTPUT" | sed 's/^.*x\([0-9]*\)$/\1/')
+  SEL_OUTPUT_VERT_PIX=$(echo "$OUTPUT_MAXRES" | grep "$SEL_OUTPUT" | sed 's/^.*x\([0-9]*\);.*$/\1/')
   if [ "$SEL_OUTPUT_VERT_PIX" -gt "1080" ]; then
-    echo "Force selected output $SEL_OUTPUT limited to FHD, disable others"
-    XRANDR_CMD=$(echo "$OUTPUT_MAXRES" | awk -v SEL_OUTPUT="$SEL_OUTPUT" 'BEGIN {FS="="; printf "xrandr"} $1==SEL_OUTPUT {printf " --output " $1 " --mode 1920x1080";} $1!=SEL_OUTPUT {printf " --output " $1 " --off";} END {printf "\n";}') 
+    echo "Force selected output $SEL_OUTPUT limited to fallback resolution, disable others"
+    XRANDR_CMD=$(echo "$OUTPUT_MAXRES" | awk -v SEL_OUTPUT="$SEL_OUTPUT" 'BEGIN {FS=";"; printf "xrandr"} $1==SEL_OUTPUT {printf " --output " $1 " --mode " $3;} $1!=SEL_OUTPUT {printf " --output " $1 " --off";} END {printf "\n";}') 
   else
     echo "Force selected output $SEL_OUTPUT, disable others"
-    XRANDR_CMD=$(echo "$OUTPUT_MAXRES" | awk -v SEL_OUTPUT="$SEL_OUTPUT" 'BEGIN {FS="="; printf "xrandr"} $1==SEL_OUTPUT {printf " --output " $1 " --auto";} $1!=SEL_OUTPUT {printf " --output " $1 " --off";} END {printf "\n";}') 
+    XRANDR_CMD=$(echo "$OUTPUT_MAXRES" | awk -v SEL_OUTPUT="$SEL_OUTPUT" 'BEGIN {FS=";"; printf "xrandr"} $1==SEL_OUTPUT {printf " --output " $1 " --auto";} $1!=SEL_OUTPUT {printf " --output " $1 " --off";} END {printf "\n";}') 
   fi
 fi